pyproject.toml
[tool.poetry]
name = "log-explorer"
version = "0.1.0"
description = "A bidirectional AI tool that translates natural language to Datadog Log Search queries"
authors = ["Carole Mendes Lago<carolemlago@gmail.com>"]
readme = "README.md"
license = "MIT"
keywords = ["datadog", "observability", "ai", "llm", "logs", "security"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.51.0"
anthropic = "^0.75.0"
datadog-api-client = "^2.46.0"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
requests = "^2.32.5"
pandas = "^2.3.3"
qdrant-client = {version = "^1.16.1", extras = ["fastembed"]}
openai = "^2.8.1"
firecrawl-py = "^4.10.0"
langchain-text-splitters = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
ruff = "^0.14.7"
mypy = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
